changed data type to decimal

+++ ByCalc/Calculators.cs
@@ -6,11 +6,10 @@ public class Calculator
 {
     private Tokenizer tokenizer = new();
     private Parser parser = new();
-    public double Calculate(string expr)
+    public decimal Calculate(string expr)
     {
         List<Token> tokens = tokenizer.Tokenize(expr);
-        /// rewrite to decimal if precision is necessary
-        double result = parser.Parse(tokens);
+        decimal result = parser.Parse(tokens);
         return result;
     }
 }
